+++ resources/todomvc/architecture-examples/preact/src/app/model.js
@@ -1,5 +1,15 @@
-import { uuid } from "./util";
+function uuid() {
+    let uuid = "";
+    for (let i = 0; i < 32; i++) {
+        let random = (Math.random() * 16) | 0;
+        // prettier-ignore
+        if (i === 8 || i === 12 || i === 16 || i === 20)
+            uuid += "-";
 
+        uuid += (i === 12 ? 4 : i === 16 ? (random & 3) | 8 : random).toString(16);
+    }
+    return uuid;
+}
 export default class TodoModel {
     constructor(key, sub) {
         this.key = key;
